+++ app/client/src/components/SearchBar/SearchBar.js
@@ -1,0 +1,11 @@
+import './SearchBar.css'
+
+export default function SearchBar({placeholder="Enter Text"}){
+  return <div id="search-bar">
+    <input type="text" placeholder={placeholder} id="search-input">
+    </input>
+    <button id="search-button">
+      <img src="/button_icons/Search_Icon.png" alt="Search" />
+    </button>
+  </div>
+}