+++ app/client/src/pages/AuthPage.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import Header from '../components/Header/Header';
 
 export default function AuthPage() {
     return (
-        <div>
+        <>
+            <Header/>
             <h1>Log In / Sign Up</h1>
             {/* TODO: replace with log in / sign up UI */}
-        </div>
+        </>
     );
 }