+++ app/client/src/pages/CategoryPage.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import Header from '../components/Header/Header';
 
 export default function CategoryPage() {
     return (
-        <div>
+        <>
+            <Header/>
             <h1>Categories</h1>
             {/* TODO: replace with category UI */}
-        </div>
+        </>
     );
 }