+++ app/client/src/pages/OrderDetailPage.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import Header from '../components/Header/Header.js';
 
 export default function OrderDetailPage() {
     return (
-        <div>
+        <>
+            <Header/>
             <h1>Order Details</h1>
             {/* TODO: replace with order details UI */}
-        </div>
+        </>
     );
 }