+++ app/client/src/pages/OrderHistoryPage.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import Header from '../components/Header/Header.js';
 
 export default function OrderHistoryPage() {
     return (
-        <div>
+        <>
+            <Header/>
             <h1>Order History</h1>
             {/* TODO: replace with order history UI */}
-        </div>
+        </>
     );
 }