+++ app/client/src/pages/ProfilePage.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import Header from '../components/Header/Header';
 
 export default function ProfilePage() {
     return (
-        <div>
+        <>
+            <Header/>
             <h1>Profile</h1>
             {/* TODO: replace with profile UI */}
-        </div>
+        </>
     );
 }