+++ app/server/controllers/CartController.js
@@ -6,6 +6,12 @@ import {
   deleteCart} 
 from "../models/cart.model.js"
 
+import {
+  addTransaction,
+  addTransactionItem
+}
+from "../models/transaction.model.js"
+
 export class CartController{
   static async update(req, res){
     const {userId, sku, count} = req.body;
@@ -39,7 +45,7 @@ export class CartController{
   }
 
   static async getAll(req, res){
-    const {userId} = req.body;
+    const {userId} = req.params;
     const result = await getCart(userId);
 
     if(result.error){
@@ -49,13 +55,26 @@ export class CartController{
   }
 
   static async getItem(req, res){
-    const {userId} = req.body;
-    const {sku} = req.params;
+    const {userId, sku} = req.params;
     const result = await getCartItem(userId, sku);
 
     if(result.error){
       return res.status(404).json({error: result.error})
     }
     return res.status(200).json(result);
+  }
+
+  static async createOrder(req, res){
+    const {userId} = req.body;
+
+    let transaction = await addTransaction(userId);
+    let cart = await getCart(userId);
+    let tid = transaction.id;
+
+    for(const item of cart){
+      await addTransactionItem(tid, item.sku, item.item_cnt);
+    }
+
+    await deleteCart(userId)
   }
 }