+++ app/server/routes/cart.routes.js
@@ -6,7 +6,8 @@ const router = Router();
 router.post('/update', CartController.update);
 router.delete('/remove/:sku', CartController.remove);
 router.delete('/clear', CartController.clear)
-router.get('/', CartController.getAll);
-router.get('/:sku', CartController.getItem);
+router.get('/:userId', CartController.getAll);
+router.get('/:userId/:sku', CartController.getItem);
+router.post('/order', CartController.createOrder);
 
 export default router;